+++ Source/Podio .NET/Models/Answer.cs
@@ -5,7 +5,7 @@ namespace PodioAPI.Models
     public class Answer
     {
         [JsonProperty("question_option_id")]
-        public int questionOptionId { get; set; }
+        public long questionOptionId { get; set; }
 
         [JsonProperty("user")]
         public User User { get; set; }
