+++ Source/Podio .NET/Models/BulkDeletionStatus.cs
@@ -9,12 +9,12 @@ namespace PodioAPI.Models
         ///     List of item ids that have been deleted at this point
         /// </summary>
         [JsonProperty(PropertyName = "deleted")]
-        public List<int> Deleted { get; set; }
+        public List<long> Deleted { get; set; }
 
         /// <summary>
         ///     List of remaining item ids to be deleted
         /// </summary>
         [JsonProperty(PropertyName = "pending")]
-        public List<int> Pending { get; set; }
+        public List<long> Pending { get; set; }
     }
 }
