+++ Source/Podio .NET/Models/Comment.cs
@@ -7,7 +7,7 @@ namespace PodioAPI.Models
     public class Comment
     {
         [JsonProperty("comment_id")]
-        public int CommentId { get; set; }
+        public long CommentId { get; set; }
 
         [JsonProperty("value")]
         public string Value { get; set; }
@@ -49,6 +49,6 @@ namespace PodioAPI.Models
         public bool IsLiked { get; set; }
 
         [JsonProperty("like_count")]
-        public int LikeCount { get; set; }
+        public long LikeCount { get; set; }
     }
 }
