+++ Source/Podio .NET/Models/FileAttachment.cs
@@ -7,7 +7,7 @@ namespace PodioAPI.Models
     public class FileAttachment
     {
         [JsonProperty("file_id")]
-        public int FileId { get; set; }
+        public long FileId { get; set; }
 
         [JsonProperty("link")]
         public string Link { get; set; }
@@ -31,7 +31,7 @@ namespace PodioAPI.Models
         public string MimeType { get; set; }
 
         [JsonProperty("size")]
-        public int? Size { get; set; }
+        public long? Size { get; set; }
 
         [JsonProperty("context")]
         public Dictionary<string, object> Context { get; set; }
