+++ Source/Podio .NET/Models/ItemField.cs
@@ -8,7 +8,7 @@ namespace PodioAPI.Models
     public class ItemField
     {
         [JsonProperty("field_id")]
-        public int? FieldId { get; set; }
+        public long? FieldId { get; set; }
 
         [JsonProperty("type")]
         public string Type { get; set; }
