+++ Source/Podio .NET/Models/PodioCollection.cs
@@ -9,13 +9,13 @@ namespace PodioAPI.Utils
         ///     Total number of items in the app
         /// </summary>
         [JsonProperty(PropertyName = "total")]
-        public int Total { get; set; }
+        public long Total { get; set; }
 
         /// <summary>
         ///     Count of items with the current filter applied
         /// </summary>
         [JsonProperty(PropertyName = "filtered")]
-        public int Filtered { get; set; }
+        public long Filtered { get; set; }
 
         [JsonProperty(PropertyName = "items")]
         public IEnumerable<T> Items { get; set; }
