+++ Source/Podio .NET/Models/QuestionAnswer.cs
@@ -5,7 +5,7 @@ namespace PodioAPI.Models
     public class QuestionAnswer
     {
         [JsonProperty("question_option_id")]
-        public int QuestionOptionId { get; set; }
+        public long QuestionOptionId { get; set; }
 
         [JsonProperty("user")]
         public Contact User { get; set; }
