+++ Source/Podio .NET/Models/QuestionOption.cs
@@ -6,7 +6,7 @@ namespace PodioAPI.Models
     public class QuestionOption
     {
         [JsonProperty("question_option_id")]
-        public int QuestionOptionId { get; set; }
+        public long QuestionOptionId { get; set; }
 
         [JsonProperty("text")]
         public String Text { get; set; }
